requirements.txt
langchain
langchain-google-genai
langchain-community
chromadb
pinecone-client
pypdf
unstructured
streamlit
python-dotenv
ragas
requests
beautifulsoup4
tiktoken
rank-bm25
psycopg2-binary>=2.9.0
sentence-transformers>=2.2.0
